backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-sqlalchemy==2.5.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
scikit-learn==0.24.2
numpy==1.21.0
openai>=1.0.0
flask-session==0.5.0 